refactor for ci upd and deployment to npm
