--- .circleci/config.yml
@@ -1,98 +1,0 @@
-version: 2.1
-commands:
-  build-js:
-    steps:
-      - run:
-          name: Js install + Build
-          command: |
-            # install
-            # node
-            curl -sL https://deb.nodesource.com/setup_14.x | sudo -E bash -
-            sudo apt-get install -y nodejs
-            # yarn
-            curl -o- -L https://yarnpkg.com/install.sh | bash
-            export PATH="$HOME/.yarn/bin:$HOME/.config/yarn/global/node_modules/.bin:$PATH"
-            # wasm-pack
-            rustup target add wasm32-unknown-unknown
-            curl https://rustwasm.github.io/wasm-pack/installer/init.sh -sSf | sh
-            # buidl
-            cd js && yarn build
-  test-js:
-    steps:
-      - run:
-          name: Js Test
-          command: |
-            # set PATH env var- not carrying over :/
-            export PATH="$HOME/.yarn/bin:$HOME/.config/yarn/global/node_modules/.bin:$PATH"
-            cd js && yarn && yarn test
-jobs:
-  build:
-    docker:
-      - image: circleci/rust:1.44.1
-    steps:
-      - checkout
-      - run:
-          name: cargo check
-          command: |
-            cargo --version --verbose
-            cargo check
-      - run:
-          name: cargo fmt
-          command: |
-            # TODO: use +nightly to enable license_template check
-            cargo fmt --version
-            cargo fmt -- --check
-      - run:
-          name: cargo test
-          command: |
-            cargo test
-      - run:
-          name: "Check 'no std' + crypto build"
-          command: |
-            # check vanilla no std
-            cargo --version --verbose
-            cargo check --no-default-features
-            # check crypto + no std
-            rustup install nightly
-            cargo +nightly --version --verbose
-            cargo +nightly check --no-default-features --features=crypto
-      # JS
-      - build-js
-      - test-js
-  publish-js:
-    docker:
-      - image: circleci/rust:1.44.1
-    steps:
-      - checkout
-      - build-js
-      - run:
-          name: Publish to NPM
-          command: |
-            echo "//registry.npmjs.org/:_authToken=${NPM_KEY}" > ~/.npmrc
-            cd js
-            npm publish --access public --tag next
-  clippy:
-    docker:
-      - image: circleci/rust:1.44.1
-    steps:
-      - checkout
-      - run:
-          name: cargo clippy
-          command: |
-            rustup component add rustfmt clippy
-            cargo clippy -- -D warnings
-workflows:
-  version: 2
-  run-build:
-    jobs:
-      - build
-      - clippy
-      # only publish to npm on release branch
-      - publish-js:
-          requires:
-          - build
-          - clippy
-          filters:
-            branches:
-              only:
-                - /^[0-9]+[.][0-9]+[.][0-9](-[a-zA-Z].+)?$/
