+++ .github/workflows/ci.yml
@@ -38,13 +38,13 @@ jobs:
     timeout-minutes: 30
     steps:
       - uses: actions/checkout@v4
-      - name: Install dependencies
-        uses: ./.github/actions/install-deps.yml
+      - name: Install & Build dependencies
+        uses: ./.github/actions/install-wasm-dependencies
       - name: cargo check wasm in js directory
         run: cargo check --target wasm32-unknown-unknown
         working-directory: js
       - name: Test JS wasm library
-        run: yarn test
+        run: pnpm test
         working-directory: js
 
   # clippy:
