--- LICENSE_TEMPLATE
@@ -1,1 +1,0 @@
-// Copyright 2022-2023 Futureverse Corporation Limited
