+++ js/.gitignore
@@ -1,5 +1,5 @@
 # js build output
 *.tgz
 node_modules/
-libNode
-libBrowser
+pkg/
+pkg-node/
