--- js/package/LICENSE
@@ -1,1 +1,0 @@
-UNLICENSED
