+++ js/test/trnnut.spec.js
@@ -1,6 +1,4 @@
-const TRNNut = require('../libNode/doughnut').TRNNut;
-// const Module = require('../libNode/doughnut').Module;
-// const Method = require('../libNode/doughnut').Method;
+const TRNNut = require('../pkg-node/doughnut').TRNNut;
 
 // The test used is same as it_works_decode_with_method_cooldown in rust
 const encodedTRNNut = new Uint8Array([
